% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2utilsVL.R
\name{areaUnderCurveDT}
\alias{areaUnderCurveDT}
\title{Returns the coordinates of a polygon for graphing the area under a curve.}
\usage{
areaUnderCurveDT(datat, startC, endC, xColumn, yColumn, yintercept = 0)
}
\arguments{
\item{datat}{the data table containing the points of the function whose area under the curve is to be plotted.}

\item{startC}{a vector containing the x and y coordinates of the leftmost point on the curve to be plotted.}

\item{endC}{a vector containing the x and y coordinates of the righttmost point on the curve to be plotted.}

\item{xColumn}{a string indicating the name of the column in datat which corresponds to the x values.}

\item{yColumn}{a string indicating the name of the column in datat which corresponds to the y values.}

\item{yintercept}{y intercept of the horizontal line at the bottom of the polygon to be plotted.}
}
\value{
a data table containing the coordinates of the polygon.
}
\description{
Returns the coordinates of a polygon for graphing the area under a curve.
}
