% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2utilsVL.R
\name{printGraTiffVL}
\alias{printGraTiffVL}
\title{Prints a ggplot2 object into a .svg and a .tiff files}
\usage{
printGraTiffVL(graphObj, fileN, width = 7, height = 7, resol = 600,
  unit = "in", compress = "LZW", showF = F,
  IMloc = "C:\\\\Program Files\\\\ImageMagick-7.0.3-Q16\\\\convert.exe",
  grayscale = F)
}
\arguments{
\item{graphObj}{ggplot2 object to print.}

\item{fileN}{Path to the files to be created. Must exclude the file extension.}

\item{width}{Width of the image in the desired measurement unit.}

\item{height}{Height of the image in the desired measurement unit.}

\item{resol}{Resolution of the image, in dpi.}

\item{unit}{Measurement unit for width and height. Defaults to inches.}

\item{compress}{compression algorithm, one of c("None", "BZip", "Fax", "Group4", "JPEG", "JPEG2000", "Lossless", "LZW", "RLE", "Zip"). Defaults to "LZW".}

\item{showF}{Boolean. Decides whether the .tiff image file should be opened after its creation in an external program. Defaults to FALSE.}

\item{IMloc}{Path to the ImageMagick executable.}

\item{grayscale}{If TRUE, the graph will be saved in grayscale (necessary for word documents)}
}
\value{
the path to the .tiff file.
}
\description{
\code{printGraTiffVL} prints a ggplot2 into  an svg and a tiff files with the desired resolution and size. Produces plot with optimal quality and with Cairo type anti-aliasing.
}
\seealso{
\url{http://www.imagemagick.org/script/command-line-options.php} \url{http://www.imagemagick.org/script/index.php}
}
